neurocuts works :)

+++ PartitionSort/stack.cpp
@@ -1,7 +1,8 @@
 #include "stack.h"
+#include <cstdint>
 
 int StackNotEmpty(stk_stack * theStack) {
-  return( theStack ? (int) theStack->top : 0);
+  return( theStack ? (intptr_t) theStack->top : 0);
 }
 
 stk_stack * StackJoin(stk_stack * stack1, stk_stack * stack2) {
